dump game & challenge information and add ability to dry run

+++ main.py
@@ -9,9 +9,10 @@ from tqdm import tqdm
 import utils.utils as utils
 
 class GzctfDumper:
-    def __init__(self, url, username, password, output_dir):
+    def __init__(self, url, dry_run, username, password, output_dir):
         self.session = self.login(url, username, password)
         self.url = url
+        self.dry_run = dry_run
         self.output_dir = output_dir
         self.games = self.get_games()
         self.game_id = self.get_game_id()
@@ -38,6 +39,7 @@ class GzctfDumper:
             print(f"An HTTP error occurred: {e}")
             if res.status_code == 401:
                 utils.print_red("Incorrect username or password")
+            self.logout()
             exit(1)
 
     def logout(self):
@@ -92,8 +94,8 @@ class GzctfDumper:
             return self.games[choice-1]['id']
 
 
-    def get_game_info(self, game_id):
-        endpoint = urljoin(self.url, f"/api/game/{game_id}")
+    def get_game_info(self):
+        endpoint = urljoin(self.url, f"/api/game/{self.game_id}")
         res = self.session.get(endpoint)
         
         try:
@@ -135,30 +137,71 @@ class GzctfDumper:
         for category, challs in self.challs.items():
             utils.print_blue(f"    [#] {category}")
             for chall in challs:
-                chall_info = f"       [-] {chall['title']} ({chall['score']})"
+                chall_info = f"        [-] {chall['title']} ({chall['score']})"
                 if chall['solved']:
                     chall_info += " [SOLVED]"
                     utils.print_green(chall_info)
                 else:
                     chall_info += " [NOT SOLVED]"
                     utils.print_red(chall_info)
+                    
+    def write_game_info(self, info, out):
+        if self.dry_run:
+            return
+
+        with open(out, 'w') as f:
+            f.write("# Game title:\n")
+            f.write(f"{info['title']}\n")
+            f.write('\n')
+
+            f.write("# Game summary:\n")
+            f.write(f"{info['summary']}\n")
+            f.write('\n')
+
+            f.write("# Game details:\n")
+            f.write(f"{info['content']}\n")
+            f.write('\n')
+
+    def dump_game_info(self):
+        info = self.get_game_info()
+        out = f"{self.output_dir}/README.md"
+
+        utils.print_yellow("[#] Downloading game information")
+        try:
+            self.write_game_info(info, out)
+            utils.print_green(f"    [-] Game information written to {out}")
+        except Exception as e:
+            utils.print_red(f"An error occurred while writing to {out}: {e}")
+
+    def write_chall_info(self, title, score, desc, hints, out):
+        if self.dry_run:
+            return
+
+        with open(out, 'w') as f:
+            f.write(f"# Title: {title}\n")
+            f.write(f"# Score: {score}\n")
+            f.write("\n")
+            
+            f.write("# Description\n")
+            f.write(f"{desc}\n")
+            f.write("\n")
+
+            f.write("# Hints\n")
+            for hint in hints:
+                f.write(f"- {hint}\n")
+
 
     def download_attachment(self, url, out, size):
+        if self.dry_run:
+            return
         with self.session.get(url, stream=True) as res:
             with tqdm(total=size, desc=out, unit='B', unit_scale=True, colour='blue', leave=False) as pb:
-                try:
-                    with open(out, 'wb') as f:
-                        res.raise_for_status()
+                with open(out, 'wb') as f:
+                    res.raise_for_status()
 
-                        for chunk in res.iter_content(chunk_size=8192):
-                            pb.update(len(chunk))
-                            f.write(chunk)
-
-                except requests.exceptions.HTTPError as e:
-                    print(f"An HTTP error occurred: {e}")
-                    self.logout()
-                    return False
-        return True
+                    for chunk in res.iter_content(chunk_size=8192):
+                        pb.update(len(chunk))
+                        f.write(chunk)
 
     def dump_challs(self):
         utils.print_yellow("[#] Downloading challenges attachments")
@@ -179,37 +222,62 @@ class GzctfDumper:
                 url = urljoin(self.url, att_url)
 
                 chall_dir = os.path.join(category_dir, title)
-                out = os.path.join(chall_dir, att_filename)
-                utils.prepare_dir(chall_dir)
+                info_out = os.path.join(chall_dir, "README.md")
+                download_out = os.path.join(chall_dir, att_filename)
 
-                if self.download_attachment(url, out, att_size):
-                    utils.print_green(f"    [-] {out} downloaded")
-                else:
-                    utils.print_red  (f"    [-] {out} download failed")
-            
+                if not self.dry_run:
+                    if not utils.is_dir_exists(chall_dir):
+                        utils.makedirs(chall_dir)
+
+                try:
+                    self.write_chall_info(title, score, desc, hints, info_out)
+                    utils.print_green(f"    [-] {info_out} written")
+                except Exception as e:
+                    print_red(f"An error occurred while writing to {out}: {e}")
+                    
+                try:
+                    self.download_attachment(url, download_out, att_size)
+                    utils.print_green(f"    [-] {download_out} downloaded")
+                except Exception as e:
+                    utils.print_red  (f"{download_out} download failed: {e}")
 
     def dump_game(self):
-        utils.prepare_dir(self.output_dir)
+        try:
+            if not self.dry_run:
+                if not utils.is_dir_exists(self.output_dir):
+                    utils.makedirs(self.output_dir)
+        except Exception as e:
+            utils.print_red(f"Unable to prepare output directory {self.output_dir}: {e}")
+            self.logout()
+            exit(1)
 
-        if not utils.is_dir_empty(self.output_dir):
-            utils.print_red("The output directory is not empty")
+        try:
+            if not utils.is_dir_empty(self.output_dir):
+                utils.print_red("The output directory is not empty")
 
-            while True:
-                choice = input("Do you want to continue? [Y/n]: ")
-                choice = choice.strip()
+                while True:
+                    choice = input("Do you want to continue? [Y/n]: ")
+                    choice = choice.strip()
 
-                if (choice == "" or
-                    choice.lower() == "y" or
-                    choice.lower() == "yes"):
-                    break
-                elif (choice.lower() == "n" or
-                    choice.lower() == "no"):
-                    return
-                else:
-                    utils.print_red("Please answer with yes/no")
+                    if (choice == "" or
+                        choice.lower() == "y" or
+                        choice.lower() == "yes"):
+                        break
+                    elif (choice.lower() == "n" or
+                        choice.lower() == "no"):
+                        self.logout()
+                        exit(1)
+                    else:
+                        utils.print_red("Please answer with yes/no")
+        except Exception as e:
+            if not self.dry_run:
+                print(f"An error occurred: {e}")
+            
 
         self.print_challs()
+        self.dump_game_info()
         self.dump_challs()
+        self.logout()
 
 
 if __name__ == "__main__":
@@ -217,11 +285,14 @@ if __name__ == "__main__":
             description="A simple python script for dumping GZ::CTF games")
 
     parser.add_argument('url',
-                        help="The base URL of GZ::CTF instance")
+                        help="Base URL of GZ::CTF instance")
+    parser.add_argument('-d', '--dry-run',
+                        action='store_true',
+                        help="Run without dumping the game")
     parser.add_argument('-u', '--username',
-                        help="The username to login with (omit for interactive username input)")
+                        help="Username to login with (omit for interactive username input)")
     parser.add_argument('-p', '--password',
-                        help="The password for the user (omit for interactive password input)")
+                        help="Password for the user (omit for interactive password input)")
     parser.add_argument('-o', '--output',
                         default="Dump",
                         help="Directory where to dump the files (default: ./Dump)")
@@ -229,6 +300,7 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     url = args.url
+    dry_run = args.dry_run
     username = args.username
     password = args.password
     output_dir = args.output
@@ -238,8 +310,7 @@ if __name__ == "__main__":
     if not args.password:
         password = getpass.getpass(prompt="Password: ")
 
-    dumper = GzctfDumper(url, username, password, output_dir)
+    dumper = GzctfDumper(url, dry_run, username, password, output_dir)
 
     dumper.dump_game()
-    dumper.logout()
